Integrate a logged in user Scene, also buttons
Added a screen for logged in user's that gives them the option to
list characters they have created, and also to create a new character.
I added a Scene for the selected character as well that displays the
name and current inventory of that character, and gives you the option
to start the game or go back. In the main game scene I added a go back
button as well as an inventory button.